+++ shared/domain/trailers/api/src/commonMain/kotlin/com.thomaskioko/tvmaniac.shared.domain.trailers.api/model/Trailer.kt
@@ -1,7 +1,7 @@
 package com.thomaskioko.tvmaniac.shared.domain.trailers.api.model
 
 data class Trailer(
-    val showId : Long,
+    val showId : Int,
     val key: String,
     val name: String,
     val youtubeThumbnailUrl: String
